rag-backend/requirements.txt
fastapi
uvicorn
python-multipart
openai>=1.0.0
faiss-cpu
numpy
PyPDF2
toml
